Motion Chair with Directional Options
A prototype of Wheelchair based on RaspberryPi is designed and complied using C. This prototype senses four switches meant for the FORWARD, REVERSE, LEFT and RIGHT movements. Based on the switch pressed, the activation of a particular motor and it's direction are determined and activated by the RaspberryPi. Though Python is easy to program for RPi, C has been chosen for it's faster execution though the libraries are not freely available. A screen shot of successful compilation has also been attached.